@@ -1,19 +1,27 @@
 package com.arctouch.codechallenge.home
 
-import android.support.v7.widget.RecyclerView
+import android.content.Intent
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.arctouch.codechallenge.R
+import com.arctouch.codechallenge.detailsScreen.DetailsScreenActivity
 import com.arctouch.codechallenge.model.Movie
 import com.arctouch.codechallenge.util.MovieImageUrlBuilder
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import kotlinx.android.synthetic.main.movie_item.view.*
 
-class HomeAdapter(private val movies: List<Movie>) : RecyclerView.Adapter<HomeAdapter.ViewHolder>() {
+class HomeAdapter : RecyclerView.Adapter<HomeAdapter.ViewHolder>() {
+
+    private var movies: MutableList<Movie> = mutableListOf()
 
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+
+        companion object {
+            const val ARG_MOVIE = "ARG_MOVIE"
+        }
 
         private val movieImageUrlBuilder = MovieImageUrlBuilder()
 
@@ -26,6 +34,12 @@ class HomeAdapter(private val movies: List<Movie>) : RecyclerView.Adapter<HomeAd
                 .load(movie.posterPath?.let { movieImageUrlBuilder.buildPosterUrl(it) })
                 .apply(RequestOptions().placeholder(R.drawable.ic_image_placeholder))
                 .into(itemView.posterImageView)
+
+            itemView.setOnClickListener {
+                val intent = Intent(it.context, DetailsScreenActivity::class.java)
+                intent.putExtra(ARG_MOVIE, movie)
+                it.context.startActivity(intent)
+            }
         }
     }
 
@@ -37,4 +51,9 @@ class HomeAdapter(private val movies: List<Movie>) : RecyclerView.Adapter<HomeAd
     override fun getItemCount() = movies.size
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) = holder.bind(movies[position])
+
+    fun addMoreMovies(moviesToAdd: List<Movie>) {
+        this.movies.addAll(moviesToAdd)
+        notifyDataSetChanged()
+    }
 }
