@@ -2,29 +2,50 @@ package com.arctouch.codechallenge.home
 
 import android.os.Bundle
 import android.view.View
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
 import com.arctouch.codechallenge.R
-import com.arctouch.codechallenge.api.TmdbApi
-import com.arctouch.codechallenge.base.BaseActivity
-import com.arctouch.codechallenge.data.Cache
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
+import com.arctouch.codechallenge.home.listener.InfinityRecyclerOnScrollListener
 import kotlinx.android.synthetic.main.home_activity.*
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class HomeActivity : BaseActivity() {
+class HomeActivity : AppCompatActivity() {
+
+    private val homeActivityViewModel: HomeActivityViewModel by viewModel()
+
+    private val homeAdapter = HomeAdapter()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.home_activity)
 
-        api.upcomingMovies(TmdbApi.API_KEY, TmdbApi.DEFAULT_LANGUAGE, 1, TmdbApi.DEFAULT_REGION)
-            .subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribe {
-                val moviesWithGenres = it.results.map { movie ->
-                    movie.copy(genres = Cache.genres.filter { movie.genreIds?.contains(it.id) == true })
-                }
-                recyclerView.adapter = HomeAdapter(moviesWithGenres)
-                progressBar.visibility = View.GONE
+        setupViews()
+        setupObservers()
+        homeActivityViewModel.getGenres()
+    }
+
+    private fun setupViews() {
+
+        recyclerView.adapter = homeAdapter
+
+        recyclerView.addOnScrollListener(object : InfinityRecyclerOnScrollListener() {
+            override fun onLoadMore() {
+                progressBarInfinityScroll.visibility = View.VISIBLE
+                homeActivityViewModel.getMovies()
             }
+        })
+    }
+
+    private fun setupObservers() {
+
+        homeActivityViewModel.genresLoaded().observe(this, Observer {
+            homeActivityViewModel.getMovies()
+        })
+
+        homeActivityViewModel.listOfMovies().observe(this, Observer {
+            homeAdapter.addMoreMovies(it)
+            progressBar.visibility = View.GONE
+            progressBarInfinityScroll.visibility = View.GONE
+        })
     }
 }
