@@ -1,10 +1,13 @@
 package com.arctouch.codechallenge.model
 
+import android.os.Parcelable
 import com.squareup.moshi.Json
+import kotlinx.android.parcel.Parcelize
 
 data class GenreResponse(val genres: List<Genre>)
 
-data class Genre(val id: Int, val name: String)
+@Parcelize
+data class Genre(val id: Int, val name: String) : Parcelable
 
 data class UpcomingMoviesResponse(
     val page: Int,
@@ -13,6 +16,7 @@ data class UpcomingMoviesResponse(
     @Json(name = "total_results") val totalResults: Int
 )
 
+@Parcelize
 data class Movie(
     val id: Int,
     val title: String,
@@ -22,4 +26,4 @@ data class Movie(
     @Json(name = "poster_path") val posterPath: String?,
     @Json(name = "backdrop_path") val backdropPath: String?,
     @Json(name = "release_date") val releaseDate: String?
-)
+) : Parcelable
